backend/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
boto3==1.34.34
botocore==1.34.34
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==3.0.4
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.1
daphne==3.0.2
Django==4.0.1
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-rest-swagger==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.5
MarkupSafe==2.1.4
msgpack==1.0.7
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
s3transfer==0.10.0
service-identity==24.1.0
shortuuid==1.0.11
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.1
